hf_dashboard/requirements.txt
streamlit
pandas
plotly
google-cloud-bigquery
python-dotenv
matplotlib